Move str_to_state to utilities
